Moves the hide related functions to a separated file dedicated for that
