+++ scanapi/utils.py
@@ -1,12 +1,4 @@
 from scanapi.errors import InvalidKeyError, MissingMandatoryKeyError
-from scanapi.settings import settings
-
-HEADERS = "headers"
-BODY = "body"
-URL = "url"
-
-ALLOWED_ATTRS_TO_HIDE = (HEADERS, BODY, URL)
-SENSITIVE_INFO_SUBSTITUTION_FLAG = "SENSITIVE_INFORMATION"
 
 
 def join_urls(first_url, second_url):
@@ -27,16 +19,6 @@ def validate_keys(keys, available_keys, required_keys, scope):
     _validate_required_keys(keys, required_keys, scope)
 
 
-def hide_sensitive_info(response):
-    report_settings = settings.get("report", {})
-    request = response.request
-    request_settings = report_settings.get("hide-request", {})
-    response_settings = report_settings.get("hide-response", {})
-
-    _hide(request, request_settings)
-    _hide(response, response_settings)
-
-
 def _validate_allowed_keys(keys, available_keys, scope):
     for key in keys:
         if not key in available_keys:
@@ -47,26 +29,3 @@ def _validate_required_keys(keys, required_keys, scope):
     if not set(required_keys) <= set(keys):
         missing_keys = set(required_keys) - set(keys)
         raise MissingMandatoryKeyError(missing_keys, scope)
-
-
-def _hide(http_msg, hide_settings):
-    for http_attr in hide_settings:
-        secret_fields = hide_settings[http_attr]
-        for field in secret_fields:
-            _override_info(http_msg, http_attr, field)
-
-
-def _override_info(http_msg, http_attr, secret_field):
-    if (
-        secret_field in getattr(http_msg, http_attr)
-        and http_attr in ALLOWED_ATTRS_TO_HIDE
-    ):
-        if http_attr == URL:
-            new_url = getattr(http_msg, http_attr).replace(
-                secret_field, SENSITIVE_INFO_SUBSTITUTION_FLAG
-            )
-            setattr(http_msg, http_attr, new_url)
-        else:
-            getattr(http_msg, http_attr)[
-                secret_field
-            ] = SENSITIVE_INFO_SUBSTITUTION_FLAG
